+++ R/OMLFlow.R
@@ -14,19 +14,7 @@
 #' `r format_bib("vanschoren2014")`
 #'
 #' @export
-#' @examples
-#' try({
-#'   library("mlr3")
-#'   # mlr3 flow:
-#'   flow = OMLFlow$new(id = 19103)
-#'   # using sugar
-#'   flow = oflw(id = 19103)
-#'   learner = as_learner(flow, "classif")
-#'   # python flow
-#'   python_flow = OMLFlow$new(19090)
-#'   # conversion to pseudo Learner
-#'   plearner = as_learner(python_flow, "classif")
-#'   }, silent = TRUE)
+#' @template examples
 OMLFlow = R6Class("OMLFlow",
   inherit = OMLObject,
   public = list(
@@ -34,25 +22,30 @@ OMLFlow = R6Class("OMLFlow",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     #'
     #' @template param_id
-    #' @template param_cache
     #' @template param_test_server
     initialize = function(
       id,
-      cache = cache_default(),
       test_server = test_server_default()
       ) {
-      super$initialize(id, cache, test_server, "flow")
+      super$initialize(id, test_server, "flow")
     },
     #' @description
     #' Prints the object.
     print = function() {
+      # trigger download first for better printing
+      self$desc
       catf("<OMLFlow:%i>", self$id)
       catf(" * Name: %s", as_short_string(self$name))
       catf(" * Dependencies: %s", paste(self$desc$dependencies, collapse = ", "))
       if (self$test_server) {
         catf(" * Using test server")
       }
-
+    },
+    #' @description
+    #' Downloads the whole object for offline usage.
+    download = function() {
+      self$desc
+      invisible(self)
     }
   ),
   active = list(