+++ R/OMLResamplingConnector.R
@@ -1,7 +1,6 @@
-OMLResamplingConnector = function(task_id = NULL, cache = cache_default(),
-  test_server = test_server_default(), parquet = parquet_default()) {
+OMLResamplingConnector = function(task_id = NULL, test_server = test_server_default(), parquet = parquet_default()) {
   if (is.null(task_id)) {
     stop(errorCondition("Argument 'task_id' must be provided", class = "missingDefaultError"))
   }
-  as_resampling(OMLTask$new(task_id, cache = cache, test_server = test_server, parquet = parquet))
+  as_resampling(OMLTask$new(task_id, test_server = test_server, parquet = parquet))
 }