+++ R/sugar.R
@@ -1,70 +1,96 @@
-#' Syntactic Sugar for Object Construction
+#' @title Syntactic Sugar for Data Construction
 #'
-#' Functions to create OpenML objects. The following functions are available:
-#' * `odt()` - creates an instance of the R6 class [`OMLData`].
-#' * `otsk()` - creates an instance of the R6 class [`OMLTask`].
-#' * `oflw()` - creates an instance of the R6 class [`OMLFlow`].
-#' * `orn()` - creates an instance of the R6 class [`OMLRun`].
-#' * `ocl()` - creates an instance of the R6 class [`OMLCollection`].
+#' @description
+#' Creates an [`OMLData`] instance.
 #'
-#' @name oml_sugar
 #' @template param_id
-#' @template param_cache
 #' @template param_test_server
 #' @template param_parquet
 #'
-#' @return ([`OMLObject`])
+#' @return ([`OMLData`])
 #'
 #' @export
 odt = function(
   id,
-  cache = cache_default(),
   parquet = parquet_default(),
   test_server = test_server_default()
   ) {
-  OMLData$new(id = id, cache = cache, parquet = parquet, test_server = test_server)
+  OMLData$new(id = id, parquet = parquet, test_server = test_server)
 }
 
-#' @rdname oml_sugar
+#' @title Syntactic Sugar for Task Construction
+#'
+#' @description
+#' Creates an [`OMLTask`] instance.
+#'
+#' @template param_id
+#' @template param_test_server
+#' @template param_parquet
+#'
+#' @return ([`OMLTask`])
+#'
 #' @export
 otsk = function(
   id,
-  cache = cache_default(),
   parquet = parquet_default(),
   test_server = test_server_default()
   ) {
-  OMLTask$new(id = id, cache = cache, parquet = parquet, test_server = test_server)
+  OMLTask$new(id = id, parquet = parquet, test_server = test_server)
 }
 
-#' @rdname oml_sugar
+#' @title Syntactic Sugar for Flow Construction
+#'
+#' @description
+#' Creates an [`OMLFlow`] instance.
+#'
+#' @template param_id
+#' @template param_test_server
+#'
+#' @return ([`OMLFlow`])
+#'
 #' @export
 oflw = function(
   id,
-  cache = cache_default(),
   test_server = test_server_default()
   ) {
-  OMLFlow$new(id = id, cache = cache, test_server = test_server)
+  OMLFlow$new(id = id, test_server = test_server)
 }
 
-#' @rdname oml_sugar
+#' @title Syntactic Sugar for Run Construction
+#'
+#' @description
+#' Creates an [`OMLRun`] instance.
+#'
+#' @template param_id
+#' @template param_test_server
+#' @template param_parquet
+#'
+#' @return ([`OMLRun`])
+#'
 #' @export
 orn = function(
   id,
-  cache = cache_default(),
   parquet = parquet_default(),
   test_server = test_server_default()
   ) {
-  OMLRun$new(id = id, cache = cache, parquet = parquet, test_server = test_server)
+  OMLRun$new(id = id, parquet = parquet, test_server = test_server)
 }
 
-#' @rdname oml_sugar
+#' @title Syntactic Sugar for Collection Construction
+#'
+#' @description
+#' Creates an [`OMLCollection`] instance.
+#'
+#' @template param_id
+#' @template param_test_server
+#'
+#' @return ([`OMLCollection`])
+#'
 #' @export
 ocl = function(
   id,
-  cache = cache_default(),
-  parquet = parquet_default(),
   test_server = test_server_default()
   ) {
-  OMLCollection$new(id = id, cache = cache, parquet = parquet, test_server = test_server)
+  OMLCollection$new(id = id, test_server = test_server)
 }