+++ man-roxygen/examples.R
@@ -1,0 +1,9 @@
+#' @examples
+#' # For technical reasons, examples cannot be included in this R package.
+#' # Instead, these are some relevant resources:
+#' #
+#' # Large-Scale Benchmarking chapter in the mlr3book:
+#' # https://mlr3book.mlr-org.com/chapters/chapter11/large-scale_benchmarking.html
+#' #
+#' # Package Article:
+#' # https://mlr3oml.mlr-org.com/articles/tutorial.html